+++ admin/lib/features/permissions/domain/usecases/delete_permission_usecase.dart
@@ -1,4 +1,5 @@
 import 'package:dartz/dartz.dart';
+import 'package:fixnum/fixnum.dart';
 import '../../../../core/errors/failures.dart';
 import '../repositories/permissions_repository.dart';
 
@@ -7,7 +8,7 @@ class DeletePermissionUseCase {
 
   DeletePermissionUseCase(this.repository);
 
-  Future<Either<Failure, void>> call(int permissionId) async {
+  Future<Either<Failure, void>> call(Int64 permissionId) async {
     return await repository.deletePermission(permissionId);
   }
 }
